feat: update api to reflect renovated backend and refactor file structure across the project
