--- src/backend/schemas/__init__.py
@@ -1,1 +1,0 @@
-from .schemas import *
